refactor(server): relocate audit helpers and remove dead stats
