+++ packages/server/package.json
@@ -19,7 +19,7 @@
   "scripts": {
     "dev": "FIRST_TREE_DEV_CALLBACK_ENABLED=1 tsx watch --env-file=../../.env src/index.ts",
     "build": "tsdown src/index.ts src/app.ts src/observability/index.ts --format esm --no-dts && node ../../scripts/copy-server-prompts.mjs",
-    "typecheck": "tsc --noEmit",
+    "typecheck": "tsc --noEmit -p tsconfig.json && tsc --noEmit -p tsconfig.scripts.json",
     "test": "vitest run",
     "coverage": "vitest run --coverage",
     "db:generate": "drizzle-kit generate && node ../../scripts/sync-migration-head.mjs",
