+++ packages/server/tsconfig.scripts.json
@@ -1,0 +1,7 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "."
+  },
+  "include": ["scripts/audit-shared-context-tree-repos.ts", "scripts/lib/**/*.ts", "scripts/__tests__/**/*.ts"]
+}
